New network_type: sum_temp_conv
Frequently MaxTempConv always focuses on the first timepoint because the
max operator breaks ties by keeping the current max (which happens to be
the first value when we pass through the ReLU and find that everything
is zero).  By summing over the temporal convolution's outputs, I'm still
collapsing the result but allowing more variation in which timepoints
are chosen to be informative.