minor fix for hooks.lua
Because we modified the naming convention for train/valid/test
accuracies that get calculated per iteration, the final output of those
quantities when we hit the max number of iterations was using the wrong
quantities.  Instead of writing the whole complicated logic to spit that
out, I just removed the final output -- we end up looking at the graphs
anyways.