Major bug fix: overlap in train/valid/test sets
I was getting weird errors (simulated data type = 2, aka no signal
simulated data, was giving above-chance accuracy when it shouldn't.
This is because there were overlapping trials in the train/valid/test
sets.  This was due to a problem with the initialization of DataFrames
by CVBySubjData and SingleSubjData.  Specifically, the trials field that
I added contained duplicate trials, so when I looked at the associated
trials for conditions like "subj_id = '02' and labels = 1", it would
sometimes spit out the same trial twice.  This was due to using
torch.linspace (which I saw used in a stack-overflow reply) instead of
torch.range.  I've modified torch.linspace everywhere.